Refactor: Member 엔티티 경로 변경 및 Member 관련 repository 구조 리팩토링
