+++ src/main/java/com/coredisc/domain/PersonalQuestion.java
@@ -2,6 +2,7 @@ package com.coredisc.domain;
 
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.mapping.QuestionCategory;
+import com.coredisc.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.*;
