+++ src/main/java/com/coredisc/domain/Post.java
@@ -4,6 +4,7 @@ package com.coredisc.domain;
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.common.enums.PostStatus;
 import com.coredisc.domain.common.enums.PublicityType;
+import com.coredisc.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.*;
