+++ src/main/java/com/coredisc/domain/TodayQuestion.java
@@ -2,10 +2,10 @@ package com.coredisc.domain;
 
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.common.enums.QuestionType;
+import com.coredisc.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
