+++ src/main/java/com/coredisc/domain/member/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.coredisc.domain.member;
+
+public interface MemberRepository {
+
+    Member save(Member member);
+    boolean existsByUsername(String username);
+    boolean existsByEmail(String email);
+    boolean existsByNickname(String nickname);
+}
