+++ src/main/java/com/coredisc/infrastructure/repository/member/JpaMemberRepository.java
@@ -1,9 +1,10 @@
 package com.coredisc.infrastructure.repository.member;
 
-import com.coredisc.domain.Member;
+import com.coredisc.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMemberRepository extends JpaRepository<Member, Long> {
+
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
     boolean existsByNickname(String nickname);
